count of shop

+++ src/main/java/com/tw/model/Shop.java
@@ -21,10 +21,10 @@ public class Shop {
 	@Column(name = "id")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
-	
+
 	@Column(name = "shop_name")
 	private String nameofshop;
-	
+
 	@Column(name = "rent")
 	private double rent;
 
@@ -38,13 +38,13 @@ public class Shop {
 	private int status;
 
 	@Column(name = "rented")
-	private double rented = 0.0;
+	private int rented = 0;
 
-	public double getRented() {
+	public int getRented() {
 		return rented;
 	}
 
-	public void setRented(double rented) {
+	public void setRented(int rented) {
 		this.rented = rented;
 	}
